% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct-distributions.R
\name{add_energy_sizeclass}
\alias{add_energy_sizeclass}
\title{Add energy & size class to raw community df}
\usage{
add_energy_sizeclass(raw_community, ln_units = 0.2)
}
\arguments{
\item{raw_community}{df of species ids and individual sizes (real, simulated, or bootstrapped)}

\item{ln_units}{natural log units to define size classes, defaults to 0.2}
}
\value{
raw_community with added columns for energy, ln_mass, size class, and the size class in g
}
\description{
Calculate energy and assign size classes
}
