% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_sampling.R
\name{boostrap_crosscomm_bseds}
\alias{boostrap_crosscomm_bseds}
\title{Get DOI for bootstrap sampled community masses from two communities}
\usage{
boostrap_crosscomm_bseds(community_a, community_b)
}
\arguments{
\item{community_df_a}{community data frame for first community}

\item{community_df_b}{community data frame for second community}
}
\value{
doi
}
\description{
Pool all individual masses from communities and randomly re-draw communities with the original number of individuals from the pool, with replacement.
Calculate the DOI of these new communities.
}
