% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_sampling.R
\name{boostrap_unif_bsed_doi}
\alias{boostrap_unif_bsed_doi}
\title{DOI of sampled community masses from uniform compared to uniform}
\usage{
boostrap_unif_bsed_doi(community_df)
}
\arguments{
\item{community_df}{Community to base samples on (nind, min and max mass)}
}
\value{
doi
}
\description{
Randomly draw N masses, where N is the total number of individuals in a community, from a uniform distribution with min and max corresponding the min and max of the entire community. 
Calculate the DOI of this BSED compared to uniform.
}
