% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_sampling.R
\name{bootstrap_crosscomm_bseds}
\alias{bootstrap_crosscomm_bseds}
\title{Get DOI for bootstrap sampled community masses from two communities}
\usage{
bootstrap_crosscomm_bseds(community_dfs)
}
\arguments{
\item{community_dfs}{list of community data frames for two communities}
}
\value{
doi
}
\description{
Pool all individual masses from communities and randomly re-draw communities with the original number of individuals from the pool, with replacement.
Calculate the DOI of these new communities.
}
