% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{calculate_bootstrap_p}
\alias{calculate_bootstrap_p}
\title{Get p value of empirical value v. bootstrapped results}
\usage{
calculate_bootstrap_p(bootstrap_dois)
}
\arguments{
\item{bootstrap_dois}{output of calculate_bootstrap_dois}
}
\value{
p value of empirical value vs. sim distribution
}
\description{
Get p value of empirical value v. bootstrapped results
}
