% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{calculate_bootstrap_uniform_dois}
\alias{calculate_bootstrap_uniform_dois}
\title{Calculate DOIs for bootstrapped BSEDs}
\usage{
calculate_bootstrap_uniform_dois(bootstrap_results)
}
\arguments{
\item{bootstrap_results}{output of draw_bootstrap_samples}
}
\value{
list of empirical DOI and sampled DOIs
}
\description{
Calculate DOIs for bootstrapped BSEDs
}
