% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-uniform.R
\name{calculate_crosscomm_dois}
\alias{calculate_crosscomm_dois}
\title{Calculate DOI for cross community bootstraps}
\usage{
calculate_crosscomm_dois(bootstrap_results)
}
\arguments{
\item{bootstrap_results}{Output of draw_bootstrap_samples}
}
\value{
list of the empirical DOI and the sampled DOIs
}
\description{
Calculate DOI for cross community bootstraps
}
