% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-uniform.R
\name{calculate_paired_dois}
\alias{calculate_paired_dois}
\title{Calculate DOIs for two community distributions}
\usage{
calculate_paired_dois(result_pair)
}
\arguments{
\item{result_pair}{List of two distributions to compare.}
}
\value{
doi
}
\description{
Just a wrapper for doi() for if the arguments are in a list
}
