% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-uniform.R
\name{calculate_uniform_size_abund_bsed}
\alias{calculate_uniform_size_abund_bsed}
\title{Calculate a BSED from a completely uniform size-abundance distribution}
\usage{
calculate_uniform_size_abund_bsed(raw_community)
}
\arguments{
\item{raw_community}{table of sizes and ids}
}
\value{
bsed from a uniform size abundance distribution of min and max size of the raw community(one individual of every size)
}
\description{
Based on a raw community df. For DOI comparisons.
}
