% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{calculate_uniform_size_energy_bsed}
\alias{calculate_uniform_size_energy_bsed}
\title{Calculate a BSED from a completely uniform size-energy distribution}
\usage{
calculate_uniform_size_energy_bsed(raw_community)
}
\arguments{
\item{raw_community}{table of sizes and ids}
}
\value{
bsed with uniform proportional energy use across size classes
}
\description{
Based on a raw community df. For DOI comparisons.
}
