% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct-distributions.R
\name{combine_communities}
\alias{combine_communities}
\title{Combine communities}
\usage{
combine_communities(indices, communities)
}
\arguments{
\item{indices}{indices of communities to pair}

\item{communities}{list of communities}
}
\value{
list of the two indices to pair
}
\description{
Helper function for setup_community_combinations
}
