% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_sampling.R
\name{community_bootstrap}
\alias{community_bootstrap}
\title{Bootstrap based on a community
Wrapper for `bootstrap_unif_bsed_doi`, possibly others.}
\usage{
community_bootstrap(community_df, bootstrap_function, nbootstraps)
}
\arguments{
\item{community_df}{Empirical community to base samples on}

\item{bootstrap_function}{Bootstrapping function to use}

\item{nbootstraps}{How many samples to draw}
}
\value{
vector of nbootstraps sampled test statistics
}
\description{
Bootstrap based on a community
Wrapper for `bootstrap_unif_bsed_doi`, possibly others.
}
