% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_ks_results.R
\name{compare_ernest_ks_values}
\alias{compare_ernest_ks_values}
\title{Compare Ernest two sample KS results to current}
\usage{
compare_ernest_ks_values(twosample_ks_results)
}
\arguments{
\item{twosample_ks_results}{result of lapply(community_combinations, twosampleks)}
}
\value{
dataframe comparing Ernest results to current results for each community
}
\description{
Compare Ernest two sample KS results to current
}
