% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_2005_data.R
\name{compare_summary_stats}
\alias{compare_summary_stats}
\title{Compare summary stats from Ernest and current data}
\usage{
compare_summary_stats(storagepath = here::here("files"))
}
\arguments{
\item{storagepath}{where the data is}
}
\value{
data frame of Ernest summary stats and current summary stats
}
\description{
Compare summary stats from Ernest and current data
}
