% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{doi}
\alias{doi}
\title{Compute distribution overlap index for BSEDs}
\usage{
doi(bsed_a, bsed_b)
}
\arguments{
\item{bsed_a}{First distribution}

\item{bsed_b}{Second distribution}
}
\value{
DOI for two distributions
}
\description{
Calculate DOI for two distributions.
}
