% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-functions.R
\name{DownloadPaperData}
\alias{DownloadPaperData}
\alias{download_raw_paper_data}
\title{Download raw paper data}
\usage{
download_raw_paper_data(from_url = FALSE,
  storage_path = here::here("working-data", "paper"))
}
\arguments{
\item{from_url}{default FALSE. If TRUE, download from source URLs. If FALSE, load from included data in repo.}

\item{storage_path}{a user-specified directory to store the data}
}
\description{
Download datasets used in Ernest 2005.
}
