% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-uniform.R
\name{draw_bootstrap_samples}
\alias{draw_bootstrap_samples}
\title{Draw bootstrap samples}
\usage{
draw_bootstrap_samples(raw_community, assumption = "uniform_size_abund",
  nbootstraps = 25)
}
\arguments{
\item{raw_community}{to base samples one}

\item{assumption}{sampling condition.}

\item{nbootstraps}{number of samples to draw}
}
\value{
list of focal bsed, sampled bseds, calculated bseds
}
\description{
Wrapper for bootstrap sampling functions
}
