% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_community_data.R
\name{draw_single_mass}
\alias{draw_single_mass}
\title{Draw a mass for an individual}
\usage{
draw_single_mass(species_id, community_masses)
}
\arguments{
\item{species_id}{the species ID (int, corresponds to element of community_masses)}

\item{community_masses}{vector of mean masses for all species in the community}
}
\value{
a mass value
}
\description{
Draw a mass for an individual given its species ID and the community masses.
}
