% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-construct-distributions.R
\name{energetic_dominance}
\alias{energetic_dominance}
\title{Calculate energetic dominance}
\usage{
energetic_dominance(community_energy, mode_cutoff = 0.05)
}
\arguments{
\item{community_energy}{community table}

\item{mode_cutoff}{defaults to 0.05, or "prop" = proportional to # of size classes in community}
}
\value{
size classes with modes and energetic dominance
}
\description{
Calculate energetic dominance
}
