% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks-test.R
\name{extract_values_twosampleks}
\alias{extract_values_twosampleks}
\title{Extract values from two sample ks test results}
\usage{
extract_values_twosampleks(this_test, val_name)
}
\arguments{
\item{this_test}{a test result}

\item{val_name}{name of the value to pull}
}
\value{
requested value
}
\description{
Helper function.
}
