% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-construct-distributions.R
\name{find_modes}
\alias{find_modes}
\title{Find energy modes from BSED}
\usage{
find_modes(bsed, mode_cutoff = 0.05)
}
\arguments{
\item{bsed}{A BSED}

\item{mode_cutoff}{defaults to 0.05, can be another number or "prop" = proportional to # of size classes in community}
}
\value{
size classes with modes
}
\description{
Find contiguous size classes with %energy > 5% of total energy. Helper for energetic dominance.
}
