% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta-corrected-ks-test.R
\name{find_ps}
\alias{find_ps}
\title{Find KS p-values}
\usage{
find_ps(rel_f, d_crits, deltaval = NULL, nvals = NULL)
}
\arguments{
\item{rel_f}{rel_f table from zar_ks_test()}

\item{d_crits}{d_crit table from zar_ks_test()}

\item{deltaval}{0, 1, or null for non-delta-corrected}

\item{nvals}{sample size, or NULL in which case i is used}
}
\value{
list of minimum and maximum p values
}
\description{
helper for ks test
}
