% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmms.R
\name{fit_gmm}
\alias{fit_gmm}
\title{Fit GMM to individual size distribution
Using `mclust::Mclust`}
\usage{
fit_gmm(isd)
}
\arguments{
\item{isd}{result of `make_isd`}
}
\value{
mclust fit
}
\description{
Fit GMM to individual size distribution
Using `mclust::Mclust`
}
