% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_sampling.R
\name{get_bootstrap_p}
\alias{get_bootstrap_p}
\title{Get p value of empirical value v. bootstrapped results}
\usage{
get_bootstrap_p(bootstrap_results)
}
\arguments{
\item{bootstrap_results}{output of community_bootstrap}
}
\value{
p value of empirical value vs. sim distribution
}
\description{
Get p value of empirical value v. bootstrapped results
}
