% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_2005_data.R
\name{get_community_name}
\alias{get_community_name}
\title{Get community name from data path}
\usage{
get_community_name(data_file)
}
\arguments{
\item{data_file}{data path}
}
\value{
community name
}
\description{
Get community name from data path
}
