% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmms.R
\name{get_gaussianmeans}
\alias{get_gaussianmeans}
\title{Get means of gaussians}
\usage{
get_gaussianmeans(gmm)
}
\arguments{
\item{gmm}{result of fit_gmm}
}
\value{
means of gaussians
}
\description{
Get means of gaussians
}
