% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_bsd.R
\name{ks_bsd}
\alias{ks_bsd}
\title{Two sample KS test for BSDs}
\usage{
ks_bsd(community_dfs, ln_mass_vals = FALSE)
}
\arguments{
\item{community_dfs}{list of dfs and names}

\item{ln_mass_vals}{whether to use the ln masses, defaults FALSE}
}
\value{
outcome list of ks_test, names
}
\description{
Two sample KS test for BSDs
}
