% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_ks_results.R
\name{load_ernest_appendixA}
\alias{load_ernest_appendixA}
\title{Load Ernest appendix A}
\usage{
load_ernest_appendixA()
}
\value{
table of results
}
\description{
Load results of d-corrected KS test
}
