% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_2005_data.R
\name{load_paper_data}
\alias{load_paper_data}
\title{Load paper community data}
\usage{
load_paper_data()
}
\arguments{
\item{datapath}{Main working directory}
}
\value{
list of data frames, one for each community.
}
\description{
Load paper community data
}
