% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-ernest-files.R
\name{load_tidy_appendix_b}
\alias{load_tidy_appendix_b}
\title{Rerrange and load table in appendix B}
\usage{
load_tidy_appendix_b()
}
\value{
table of site comparisons, and d and p values, for KS tests.
}
\description{
Rerrange and load table in appendix B
}
