% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_distributions.R
\name{make_bsd}
\alias{make_bsd}
\title{Construct BSD from df with community data}
\usage{
make_bsd(community_df, decimals = 2, ln_units = 0.2)
}
\arguments{
\item{decimals}{how many decimals}

\item{ln_units}{defaults to 0.2}

\item{community_table}{df of species ids, individual sizes, energy, and size class}
}
\value{
bsd of species and mean masses in g and log(mean mass)
}
\description{
species level body size distributions
}
