% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct-distributions.R
\name{make_bsed}
\alias{make_bsed}
\title{Construct BSED from df with size classes}
\usage{
make_bsed(community_energy, decimals = NULL)
}
\arguments{
\item{community_energy}{df of species ids, individual sizes, energy, and size class}

\item{decimals}{default NULL; decimals to round size classes}
}
\value{
bsed
}
\description{
Calculate total energy of individuals per size class
}
