% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_distributions.R,
%   R/construct_distributions.R
\name{make_community_table}
\alias{make_community_table}
\title{Add energy & size class to raw community df}
\usage{
make_community_table(community, ln_units = 0.2)

make_community_table(community, ln_units = 0.2)
}
\arguments{
\item{ln_units}{natural log units to define size classes, defaults to 0.2}

\item{community_df}{df of species ids and individual sizes (real, simulated, or bootstrapped)}

\item{community_df}{df of species ids and individual sizes (real, simulated, or bootstrapped)}

\item{ln_units}{natural log units to define size classes, defaults to 0.2}
}
\value{
expanded community_df

expanded community_df
}
\description{
Calculate energy and assign size classes

Calculate energy and assign size classes
}
