% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_2005_data.R
\name{make_ernest_name_key}
\alias{make_ernest_name_key}
\title{Make key to Ernest names - current community names
Ernest (2005) refers to the communities with the `site` name in the text and Figure 1 To compare the communities above to the communities in the resurrected data set, we can try to match them based on the BSD and BSED plots and species richness.}
\usage{
make_ernest_name_key()
}
\description{
Make key to Ernest names - current community names
Ernest (2005) refers to the communities with the `site` name in the text and Figure 1 To compare the communities above to the communities in the resurrected data set, we can try to match them based on the BSD and BSED plots and species richness.
}
