% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmms.R
\name{make_isd}
\alias{make_isd}
\title{Make ISD}
\usage{
make_isd(community_energy)
}
\arguments{
\item{community_energy}{community df with energy & sizeclass columns added}
}
\value{
ISD table
}
\description{
Make ISD
}
