% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_bootstrap_pvals}
\alias{plot_bootstrap_pvals}
\title{Plot histogram of p values}
\usage{
plot_bootstrap_pvals(bootstrap_dois)
}
\arguments{
\item{bootstrap_dois}{list of dois for empirical, sampled distributions}
}
\value{
histogram of p values
}
\description{
Plot histogram of p values
}
