% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distributions.R
\name{plot_bsd}
\alias{plot_bsd}
\title{Plot a single BSD}
\usage{
plot_bsd(bsd, bsd_name = NULL)
}
\arguments{
\item{bsd}{body size distribution table}

\item{bsd_name}{community name}
}
\value{
density plot of avg. size x nspeices
}
\description{
Plot a single BSD
}
