% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distributions.R
\name{plot_bsed}
\alias{plot_bsed}
\title{Plot a single BSED}
\usage{
plot_bsed(bsed, bsed_name = NULL)
}
\arguments{
\item{bsed}{body size energy distribution table}

\item{bsed_name}{community name}
}
\value{
barplot of proportion energy use by size class
}
\description{
Plot a single BSED
}
