% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_bsed_bootstrap_results}
\alias{plot_bsed_bootstrap_results}
\title{Plot a single BSED bootstrapping result}
\usage{
plot_bsed_bootstrap_results(bootstrap_dois, bsed_name = NULL)
}
\arguments{
\item{bootstrap_dois}{list of empirical DOI and sampled DOIs}

\item{bsed_name}{community name, optional}
}
\value{
histogram of bootstrapped DOI values, line for empirical value.
}
\description{
Plot a single BSED bootstrapping result
}
