% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_crosscomm_bseds}
\alias{plot_crosscomm_bseds}
\title{Plot cross community comparisons in one panel}
\usage{
plot_crosscomm_bseds(cross_community_comparisons)
}
\arguments{
\item{cross_community_comparisons}{all comparisons}
}
\value{
comparison plots in a single plot
}
\description{
Plot cross community comparisons in one panel
}
