% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_crosscomm_ks_pvals}
\alias{plot_crosscomm_ks_pvals}
\title{Plot comparison of Ernest vs current KS test p values}
\usage{
plot_crosscomm_ks_pvals(results_comparison)
}
\arguments{
\item{results_comparison}{result of compare_ernest_ks_values}
}
\value{
histogram of Ernest and current p values
}
\description{
Plot comparison of Ernest vs current KS test p values
}
