% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_e_dom}
\alias{plot_e_dom}
\title{Plot energetic dominance}
\usage{
plot_e_dom(e_doms)
}
\arguments{
\item{energetic_dominances}{either a single data table of $E_D$ or a list of data tables. If a list, will be combined into a single data table.}
}
\value{
histogram of $E_D$ values
}
\description{
Generate histogram of energetic dominance values
}
