% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmms.R
\name{plot_gmm}
\alias{plot_gmm}
\title{Density plot of GMM}
\usage{
plot_gmm(gmm, gmm_name = NULL)
}
\arguments{
\item{gmm}{result of fit_gmm}

\item{gmm_name}{not used}
}
\value{
density plot
}
\description{
Density plot of GMM
}
