% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmms.R
\name{plot_isd}
\alias{plot_isd}
\title{Plot isd}
\usage{
plot_isd(isd, isd_name = NULL)
}
\arguments{
\item{isd}{result of `make_isd`}

\item{isd_name}{community name, defaults NULL}
}
\value{
plot of ISD
}
\description{
Plot isd
}
