% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distributions.R
\name{plot_paper_bseds}
\alias{plot_paper_bseds}
\title{Replicate Figure 1}
\usage{
plot_paper_bseds(bseds)
}
\arguments{
\item{bseds}{list of bseds}
}
\value{
9 panel plot of bseds modeled after Ernest 2005 Figure 1
}
\description{
Replicate Figure 1
}
