% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distributions.R
\name{plot_paper_dists}
\alias{plot_paper_dists}
\title{Replicate Figure 1}
\usage{
plot_paper_dists(dists, dist_type)
}
\arguments{
\item{dists}{list of bseds or bsds}

\item{dist_type}{'bsed', 'bsd', 'both' ('both' not yet functional)}
}
\value{
9 panel plot of bseds/bsds/both modeled after Ernest 2005 Figure 1
}
\description{
Replicate Figure 1
}
