% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-functions.R
\name{ProcessAllRawData}
\alias{ProcessAllRawData}
\alias{process_raw_data}
\title{Process raw data in to the appropriate format.}
\usage{
process_raw_data(storage_path = here::here("working-data", "paper"))
}
\arguments{
\item{storage_path}{where to get and put the data}
}
\description{
Process all the datasets and make the key of site names
}
