% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{ProcessAndrews}
\alias{ProcessAndrews}
\alias{process_andrews_data}
\title{Process Andrews data in to the appropriate format.}
\usage{
process_andrews_data(storage_path = here::here("working-data", "paper"))
}
\arguments{
\item{storage_path}{where to get and put the data}
}
\description{
Process Andrews smammal data
}
