% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-functions.R
\name{ProcessNiwot}
\alias{ProcessNiwot}
\alias{process_niwot_data}
\title{Process Niwot data into the appropriate format.}
\usage{
process_niwot_data(storage_path = here::here("working-data", "paper"))
}
\arguments{
\item{storage_path}{where to put the data}
}
\description{
Process Niwot smammal data
Initial cleaning from LTER website code generator
}
