% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-functions.R
\name{ProcessPortal}
\alias{ProcessPortal}
\alias{process_portal_data}
\title{Process Portal data in to the appropriate format.}
\usage{
process_portal_data(storage_path = here::here("working-data", "paper"),
  portaldatapath = NULL)
}
\arguments{
\item{storage_path}{where to put the data}

\item{portaldatapath}{where the portal data is}
}
\description{
Process Portal smammal data
Using `portalr`
}
