% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-functions.R
\name{ProcessSev}
\alias{ProcessSev}
\alias{process_sev_data}
\title{Process Sevilleta data in to the appropriate format.}
\usage{
process_sev_data(storage_path = here::here("working-data", "paper"))
}
\arguments{
\item{storage_path}{where to put the data}
}
\description{
Process Sevilleta smammal data
Initial cleaning from LTER website code generator
}
