% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-uniform.R
\name{sample_cross_communities_bsed}
\alias{sample_cross_communities_bsed}
\title{Bootstrap compare two communities}
\usage{
sample_cross_communities_bsed(community_pair)
}
\arguments{
\item{community_pair}{list of 2 communities to compare}
}
\value{
list of 2 bseds
}
\description{
Randomly re-draw communities with masses and nindividuals of original communities.
}
