% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-uniform.R
\name{sample_uniform_size_abund_bsed}
\alias{sample_uniform_size_abund_bsed}
\title{Draw a sample community with a uniform size-abundance distribution}
\usage{
sample_uniform_size_abund_bsed(raw_community)
}
\arguments{
\item{raw_community}{Community to base new community on.}
}
\value{
bsed of sampled community
}
\description{
With the same minimum and maximum body size as an empirical community.
}
