% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{sample_uniform_size_energy_bsed}
\alias{sample_uniform_size_energy_bsed}
\title{Draw a sample community with a uniform size-energy distribution}
\usage{
sample_uniform_size_energy_bsed(raw_community)
}
\arguments{
\item{raw_community}{Community to base new community on.}
}
\value{
bsed of sampled community
}
\description{
Uniform energy distribution across size classes
}
