% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct-distributions.R
\name{setup_community_combinations}
\alias{setup_community_combinations}
\title{Set up pairwise community combinations}
\usage{
setup_community_combinations(communities)
}
\arguments{
\item{communities}{list of communities to compare to each other}
}
\value{
list of all possible combinations of two communities.
}
\description{
Set up a list of lists of all possible 2-way comparisons of communities in a list of communities.
}
