% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_community_data.R
\name{simulate_community}
\alias{simulate_community}
\title{Simulate an entire community}
\usage{
simulate_community(nspecies, nind, min_mass, max_mass)
}
\arguments{
\item{nspecies}{S}

\item{nind}{N}

\item{min_mass}{Min average mass}

\item{max_mass}{Max average mass}
}
\value{
matrix with 1 row per individual and two columns: species assignments and mass.
}
\description{
Randomly generate masses for N individuals randomly distributed among S species assignments & with minimum and maximum body sizes.
}
