% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_community_data.R
\name{simulate_ind_mass}
\alias{simulate_ind_mass}
\title{Simulate masses for individuals}
\usage{
simulate_ind_mass(individual_species, community_masses)
}
\arguments{
\item{individual_species}{Vector of species identities for each individual}

\item{community_masses}{Vector of mean body size for each species}
}
\value{
vector of masses for N individuals, drawn according to the species' mean mass and the species assignments.
}
\description{
Randomly generate masses for individuals given species assignments & mean masses for each species. Assumes a normal distribution with SD = 0.1 * mass.
}
