% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_community_data.R
\name{simulate_ind_species}
\alias{simulate_ind_species}
\title{Simulate species identities for individuals}
\usage{
simulate_ind_species(nspecies, nind)
}
\arguments{
\item{nspecies}{S, number of species}

\item{nind}{N, number of individuals}
}
\value{
vector of species identities (1:S) for N individuals
}
\description{
Randomly assign species identities from 1 to S to N individuals.
}
