% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_2005_data.R
\name{tidy_appendix_b}
\alias{tidy_appendix_b}
\title{Rerrange table in appendix B}
\usage{
tidy_appendix_b(storagepath = "files")
}
\arguments{
\item{storagepath}{where the data is}
}
\value{
table of site comparisons, and d and p values, for KS tests.
}
\description{
Rerrange table in appendix B
}
