% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_sampling.R
\name{uniform_size_energy_bsed}
\alias{uniform_size_energy_bsed}
\title{Uniform size-energy BSED}
\usage{
uniform_size_energy_bsed(community_df)
}
\arguments{
\item{community_df}{table of sizes and ids}
}
\value{
bsed drawn from a uniform body size energy distribution
}
\description{
Based on a community_df
}
