% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_corrected_ks.R
\name{zar_ks_test}
\alias{zar_ks_test}
\title{d-Corrected Kolgomorov-Smirnov test}
\usage{
zar_ks_test(bsd, delta_correction = F)
}
\arguments{
\item{bsd}{distribution}

\item{delta_correction}{T/F use delta correction for small sample sizes}
}
\value{
list of signif to p = 0.05, d value, dcrit
}
\description{
delta-corrected KS test
}
