% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest-files.R
\name{compare_summary_stats}
\alias{compare_summary_stats}
\title{Compare summary stats from Ernest and current data}
\usage{
compare_summary_stats(storage_path = here::here("working-data", "paper",
  "processed"))
}
\arguments{
\item{storage_path}{where the user accessible data is}
}
\value{
data frame of Ernest summary stats and current summary stats
}
\description{
Compare summary stats from Ernest and current data
}
