% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_corrected_ks.R, R/doi.R
\name{doi}
\alias{doi}
\title{d-Corrected Kolgomorov-Smirnov test}
\usage{
doi(dist_a, dist_b = NULL)

doi(dist_a, dist_b = NULL)
}
\arguments{
\item{dist_a}{First distribution}

\item{dist_b}{Second distribution, if NULL, uniform}

\item{dist_a}{First distribution}

\item{dist_b}{Second distribution, if NULL, uniform}
}
\value{
KS statistic

DOI for two distributions
}
\description{
$delta$-corrected KS test

Calculate DOI for two distributions or one distribution compared to uniform.
}
