% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_corrected_ks.R, R/doi.R
\name{doi}
\alias{doi}
\title{d-Corrected Kolgomorov-Smirnov test}
\usage{
doi(bsed_a, bsed_b)

doi(bsed_a, bsed_b)
}
\arguments{
\item{bsed_a}{First distribution}

\item{bsed_b}{Second distribution}

\item{dist_a}{First distribution}

\item{dist_b}{Second distribution, if NULL, uniform}
}
\value{
KS statistic

DOI for two distributions
}
\description{
$delta$-corrected KS test

Calculate DOI for two distributions.
}
