% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_2005_data.R
\name{DownloadPaperData}
\alias{DownloadPaperData}
\alias{download_raw_paper_data}
\title{Download raw paper data}
\usage{
download_raw_paper_data(datapath = here::here("data", "paper", "raw"))
}
\arguments{
\item{datapath}{Folder to store the data}
}
\description{
Download datasets used in Ernest 2005.
Only downloads if the relevant folders do not exist in the `datapath`.
}
