% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_distributions.R
\name{energetic_dominance}
\alias{energetic_dominance}
\title{Calculate energetic dominance}
\usage{
energetic_dominance(community_df)
}
\arguments{
\item{community_df}{}
}
\value{
size classes with modes and energetic dominance
}
\description{
Calculate energetic dominance
}
