% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_distributions.R
\name{find_modes}
\alias{find_modes}
\title{Find energy modes from BSED}
\usage{
find_modes(bsed)
}
\arguments{
\item{bsed}{A BSED}
}
\value{
size classes with modes
}
\description{
Find contiguous size classes with %energy > 5% of total energy
}
