% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_corrected_ks.R
\name{find_ps}
\alias{find_ps}
\title{find ps}
\usage{
find_ps(rel_f, d_crits, deltaval = NULL, nvals = NULL)
}
\arguments{
\item{rel_f}{rel_f table}

\item{d_crits}{d_crit table}

\item{deltaval}{0, 1, or null for non-delta-corrected}

\item{nvals}{sample size, or NULL in which case i is used}
}
\value{
list of pvals
}
\description{
helper for ks test
}
