% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks-test.R
\name{ks_bsd}
\alias{ks_bsd}
\title{Two sample KS test for BSDs}
\usage{
ks_bsd(raw_communities, ln_mass_vals = FALSE)
}
\arguments{
\item{raw_communities}{list of two dfs and names}

\item{ln_mass_vals}{whether to use the ln masses, defaults FALSE}
}
\value{
outcome list of ks_test, names of communities
}
\description{
Two sample KS test for BSDs
}
