% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functions.R
\name{load_paper_data}
\alias{load_paper_data}
\title{Load paper community data}
\usage{
load_paper_data(storage_path = here::here("working-data", "paper",
  "processed"))
}
\arguments{
\item{storage_path}{Main working directory}
}
\value{
list of data frames, one for each community.
}
\description{
Load paper community data
}
