% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct-distributions.R
\name{make_bsd}
\alias{make_bsd}
\title{Construct BSD from df with community data}
\usage{
make_bsd(raw_community, decimals = NULL, ln_units = 0.2)
}
\arguments{
\item{raw_community}{df of species ids and individual sizes}

\item{decimals}{how many decimals for rounding, defaults to NULL}

\item{ln_units}{defaults to 0.2}
}
\value{
bsd of species and mean masses in g and log(mean mass)
}
\description{
species level body size distributions
}
