% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_distributions.R
\name{make_bsed}
\alias{make_bsed}
\title{Construct BSED from df with size classes}
\usage{
make_bsed(community_df, decimals = NULL)
}
\arguments{
\item{community_df}{df of species ids, individual sizes, energy, and size class}
}
\value{
bsed
}
\description{
Calculate total energy of individuals per size class
}
