% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bootstrap_results.R
\name{plot_bootstrap_pvals}
\alias{plot_bootstrap_pvals}
\title{Plot histogram of p values}
\usage{
plot_bootstrap_pvals(bootstrap_result)
}
\arguments{
\item{bootstrap_results}{results objects}
}
\value{
histogram of p values
}
\description{
Plot histogram of p values
}
