% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_bsd.R
\name{plot_crosscomm_ks_pvals}
\alias{plot_crosscomm_ks_pvals}
\title{Plot histogram of p values}
\usage{
plot_crosscomm_ks_pvals(crosscomm_result)
}
\arguments{
\item{crosscomm_result}{result df}
}
\value{
histogram of p values
}
\description{
Plot histogram of p values
}
