% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_paper_dists}
\alias{plot_paper_dists}
\title{Replicate Figure 1}
\usage{
plot_paper_dists(dists, dist_type)
}
\arguments{
\item{dists}{list of bseds or bsds, or P values for BSED bootstrap analysis}

\item{dist_type}{"bsed", "bsd", or "bsed_bootstraps"}
}
\value{
9 panel plot modeled after Ernest 2005 Figure 1
}
\description{
plot all 9 BSEDs or BSDs in the same arrangement as Ernest Figure 1
}
