% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_2005_data.R
\name{ProcessAndrews}
\alias{ProcessAndrews}
\alias{process_andrews_data}
\title{Process Andrews data in to the appropriate format.}
\usage{
process_andrews_data(datapath = here::here())
}
\description{
Process Andrews smammal data
#'
}
