% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_2005_data.R
\name{ProcessNiwot}
\alias{ProcessNiwot}
\alias{process_niwot_data}
\title{Process Niwot data into the appropriate format.}
\usage{
process_niwot_data(datapath = here::here())
}
\arguments{
\item{datapath}{path to where data is}
}
\description{
Process Niwot smammal data
Initial cleaning from LTER website code generator
}
