% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_2005_data.R
\name{ProcessPortal}
\alias{ProcessPortal}
\alias{process_portal_data}
\title{Process Portal data in to the appropriate format.}
\usage{
process_portal_data(storagepath = "files",
  portaldatapath = "/Users/renatadiaz/Documents/GitHub/weecology/")
}
\arguments{
\item{storagepath}{where to put the data}

\item{portaldatapth}{where the portal data is}
}
\description{
Process Portal smammal data
Using `portalr`
}
