% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_2005_data.R
\name{ProcessSev}
\alias{ProcessSev}
\alias{process_sev_data}
\title{Process Sevilleta data in to the appropriate format.}
\usage{
process_sev_data(datapath = here::here())
}
\description{
Process Sevilleta smammal data
Initial cleaning from LTER website code generator
}
