% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_sampling.R
\name{uniform_size_abund_bsed}
\alias{uniform_size_abund_bsed}
\title{Uniform size-abundance BSED}
\usage{
uniform_size_abund_bsed(community_df)
}
\arguments{
\item{community_df}{table of sizes and ids}
}
\value{
bsed drawn from a uniform size abundance distribution (one individual of every size)
}
\description{
Based on a community_df
}
