% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_corrected_ks.R
\name{zar_ks_test}
\alias{zar_ks_test}
\title{d-Corrected Kolgomorov-Smirnov test}
\usage{
zar_ks_test(distribution, delta_correction = F,
  focal_column = "species_mean_mass", expected_range = NULL,
  n_or_i = "n", storagepath = here::here("files"))
}
\arguments{
\item{delta_correction}{T/F use delta correction for small sample sizes}

\item{focal_column}{name of column for the distribution to be evaluated}

\item{expected_range}{vector of expected min and max values for uniform; defaults to min and max of measurement column}

\item{n_or_i}{"n" or "i"}

\item{storagepath}{where files are}

\item{bsd}{distribution}
}
\value{
list of signif to p = 0.05, d value, dcrit
}
\description{
delta-corrected KS test
}
